SRM_TRP_Chatbot/backend/requirements.txt
langchain_core
langchain
python-dotenv==1.0.1
langchain_community
langserve
bs4==0.0.2
pypdf==4.3.1
groq==0.9.0
cassio== 0.1.8
beautifulsoup4==4.12.3
langchain-groq
PyPDF2==3.0.1
langchain-objectbox
flask==3.0.3
flask-cors==4.0.1
ragstack-ai
langchain-nvidia-ai-endpoints
datasets
gunicorn